requirements/requirements-test.txt
flake8==3.8.4
isort==5.7.0
mypy==0.790
pytest==6.2.1
pytest-cov>=2.10.1
pytest-django>=4.1.0
pyyaml==5.3.1
twine==3.3.0
